+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add initial support for encoding, decoding, representing and manipulating the `trade` entity.
 * Add `SystemConfig.isLinked(String)` helper method.
 * Separate the representation of `minables` and `asteroids` which involved; introducing a separate config object `SystemConfig.Minable`, modifying existing methods on `SystemConfig` that manipulate `minables`, introducing new methods to access `minables` and removing the `SystemConfig.Asteroid.isMinable()` method.
+* Refactor several of the `SystemConfig.addX(...)` methods to return the object created.
 
 ### [v0.03](https://github.com/realityforge/zifnab/tree/v0.03) (2019-08-10)
 [Full Changelog](https://github.com/realityforge/zifnab/compare/v0.02...v0.03)